refactor dan validators into a separated module
