+++ base_layer/core/src/validation/mod.rs
@@ -43,6 +43,7 @@ pub use traits::{
 };
 
 pub mod block_validators;
+mod dan_validators;
 mod difficulty_calculator;
 pub use difficulty_calculator::*;
 pub mod header_validator;
