+++ base_layer/core/src/validation/test.rs
@@ -283,37 +283,7 @@ fn chain_balance_validation() {
 }
 
 mod transaction_validator {
-    use std::convert::TryInto;
-
-    use tari_common_types::types::{FixedHash, PublicKey, Signature};
-    use tari_utilities::hex::Hex;
-
     use super::*;
-    use crate::{
-        block_spec,
-        test_helpers::blockchain::TestBlockchain,
-        transactions::{
-            tari_amount::T,
-            test_helpers::{spend_utxos, TransactionSchema},
-            transaction_components::{
-                vec_into_fixed_string,
-                CheckpointParameters,
-                CommitteeMembers,
-                ConstitutionChangeFlags,
-                ConstitutionChangeRules,
-                ContractAcceptanceRequirements,
-                ContractConstitution,
-                ContractDefinition,
-                ContractSpecification,
-                RequirementsForConstitutionChange,
-                SideChainConsensus,
-                Transaction,
-                UnblindedOutput,
-            },
-        },
-        txn_schema,
-        validation::transaction_validators::TxConsensusValidator,
-    };
 
     #[test]
     fn it_rejects_coinbase_outputs() {
@@ -325,125 +295,5 @@ mod transaction_validator {
         let (tx, _, _) = tx!(MicroTari(100_000), fee: MicroTari(5), inputs: 1, outputs: 1, features: features);
         let err = validator.validate(&tx).unwrap_err();
         unpack_enum!(ValidationError::ErroneousCoinbaseOutput = err);
-    }
-
-    pub fn schema_to_transaction(txns: &[TransactionSchema]) -> (Vec<Arc<Transaction>>, Vec<UnblindedOutput>) {
-        let mut tx = Vec::new();
-        let mut utxos = Vec::new();
-        txns.iter().for_each(|schema| {
-            let (txn, mut output) = spend_utxos(schema.clone());
-            tx.push(Arc::new(txn));
-            utxos.append(&mut output);
-        });
-        (tx, utxos)
-    }
-
-    pub fn create_block(
-        blockchain: &mut TestBlockchain,
-        block_name: &'static str,
-        schema: TransactionSchema,
-    ) -> Vec<UnblindedOutput> {
-        let (txs, outputs) = schema_to_transaction(&[schema]);
-        let (_, _) = blockchain
-            .append_to_tip(block_spec!(block_name, transactions: txs.iter().map(|t| (**t).clone()).collect()))
-            .unwrap();
-
-        outputs
-    }
-
-    fn create_contract_definition_schema(input: UnblindedOutput) -> (FixedHash, TransactionSchema) {
-        let definition = ContractDefinition {
-            contract_name: vec_into_fixed_string("name".as_bytes().to_vec()),
-            contract_issuer: PublicKey::default(),
-            contract_spec: ContractSpecification {
-                runtime: vec_into_fixed_string("runtime".as_bytes().to_vec()),
-                public_functions: vec![],
-            },
-        };
-        let contract_id = definition.calculate_contract_id();
-        let definition_features = OutputFeatures::for_contract_definition(definition);
-
-        let tx_schema =
-            txn_schema!(from: vec![input], to: vec![0.into()], fee: 5.into(), lock: 0, features: definition_features);
-
-        (contract_id, tx_schema)
-    }
-
-    fn create_contract_constitution_schema(contract_id: FixedHash, input: UnblindedOutput) -> TransactionSchema {
-        let validator_committee: CommitteeMembers = vec![PublicKey::default()].try_into().unwrap();
-        let constitution = ContractConstitution {
-            validator_committee,
-            acceptance_requirements: ContractAcceptanceRequirements {
-                acceptance_period_expiry: 100,
-                minimum_quorum_required: 5,
-            },
-            consensus: SideChainConsensus::MerkleRoot,
-            checkpoint_params: CheckpointParameters {
-                minimum_quorum_required: 5,
-                abandoned_interval: 100,
-            },
-            constitution_change_rules: ConstitutionChangeRules {
-                change_flags: ConstitutionChangeFlags::all(),
-                requirements_for_constitution_change: Some(RequirementsForConstitutionChange {
-                    minimum_constitution_committee_signatures: 5,
-                    constitution_committee: Some(
-                        vec![PublicKey::default(); CommitteeMembers::MAX_MEMBERS]
-                            .try_into()
-                            .unwrap(),
-                    ),
-                }),
-            },
-            initial_reward: 100.into(),
-        };
-        let constitution_features = OutputFeatures::for_contract_constitution(contract_id, constitution);
-
-        txn_schema!(from: vec![input], to: vec![0.into()], fee: 5.into(), lock: 0, features: constitution_features)
-    }
-
-    fn create_contract_acceptance_schema(contract_id: FixedHash, input: UnblindedOutput) -> TransactionSchema {
-        // let validator_node_public_key = PublicKey::default();
-        let validator_node_public_key =
-            PublicKey::from_hex("70350e09c474809209824c6e6888707b7dd09959aa227343b5106382b856f73a").unwrap();
-        let signature = Signature::default();
-
-        let acceptance_features =
-            OutputFeatures::for_contract_acceptance(contract_id, validator_node_public_key, signature);
-
-        let mut tx =
-            txn_schema!(from: vec![input], to: vec![0.into()], fee: 5.into(), lock: 0, features: acceptance_features);
-        tx.output_version = None;
-
-        tx
-    }
-
-    #[test]
-    fn it_rejects_contract_acceptances_of_non_committee_member() {
-        let consensus_manager = ConsensusManagerBuilder::new(Network::LocalNet).build();
-        let mut blockchain = TestBlockchain::create(consensus_manager);
-        let (_, coinbase_a) = blockchain.add_next_tip(block_spec!("1")).unwrap();
-
-        let schema = txn_schema!(from: vec![coinbase_a], to: vec![50 * T, 50 * T, 50 * T]);
-        let change_outputs = create_block(&mut blockchain, "2", schema);
-
-        let (contract_id, schema) = create_contract_definition_schema(change_outputs[0].clone());
-        // let schema = txn_schema!(from: vec![change_outputs[0].clone()], to: vec![10 * T]);
-        create_block(&mut blockchain, "3", schema);
-
-        // let schema = txn_schema!(from: vec![change_outputs[1].clone()], to: vec![10 * T]);
-        let schema = create_contract_constitution_schema(contract_id, change_outputs[1].clone());
-        create_block(&mut blockchain, "4", schema);
-
-        let schema = create_contract_acceptance_schema(contract_id, change_outputs[2].clone());
-        let (txs, _) = schema_to_transaction(&[schema]);
-
-        let validator = TxConsensusValidator::new(blockchain.db().clone());
-        let err = validator.validate(txs.first().unwrap()).unwrap_err();
-
-        match err {
-            ValidationError::ConsensusError(message) => {
-                assert!(message.contains("Invalid contract acceptance: validator node public key is not in committee"))
-            },
-            _ => panic!("Expected a consensus error"),
-        }
     }
 }
